requirements.txt
rich>=10.0.0
inquirer>=3.1.3
pyyaml>=6.0.1
openai>=1.12.0
python-dotenv>=1.0.0
anthropic>=0.8.0
playwright>=1.42.0
pytest>=7.4.0
pytest-cov>=4.1.0
schedule==1.2.0



aider-chat
python-dotenv
